CCHub fix

+++ BlueSandsLMS.Common/DTOs/Admin/GlobalAdminDtos.cs
@@ -15,7 +15,7 @@ int totalStudent,
 int ActiveSubscriptions,
 
 
-int TotalPayments,
+decimal TotalPayments,
 int TotalStemCourses,
 double TotalQuizScores,
 int TotalSubscribedUsers,
